Move read_url to iterm2 module
Only iterm2 actually uses it.

Fixes unused warning when compiling only with terminology.
